# docs/swagger.yaml
definitions:
...
  handlers.AuthRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.CourseRequest:
    properties:
      title:
        type: string
    type: object
...
  handlers.UpdateRequest:
    properties:
      title:
        type: string
    type: object
...
  models.ApiError:
    properties:
      error:
        type: string
    type: object
  models.Course:
    properties:
      id:
        type: string
      title:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        example: error description
        type: string
    type: object
  models.LoginResponse:
    properties:
      expires:
        example: 1672531200
        type: integer
      role:
        example: admin
        type: string
      token:
        example: eyJhbGciOi...
        type: string
    type: object
  models.MessageResponse:
    properties:
      message:
        example: success message
        type: string
    type: object
...
  models.TokenResponse:
    properties:
      expires:
        example: 1672531200
        type: integer
      token:
        example: eyJhbGciOi...
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Вход в систему с email и паролем
      parameters:
      - description: Данные для входа
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Аутентификация пользователя
      tags:
      - Auth
  /auth/logout:
    post:
      description: Завершает текущую сессию пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Выход из системы
      tags:
      - Auth
...
  /auth/refresh:
    post:
      description: Обновляет JWT токен с помощью refresh токена из cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Обновление токена
      tags:
      - Auth
...
  /settings/courses:
    get:
      description: Возвращает список всех курсов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Course'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Получить все курсы
      tags:
      - Courses
    post:
      consumes:
      - application/json
      description: Создает новый курс
      parameters:
      - description: Данные курса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Создать курс
      tags:
      - Courses
  /settings/courses/{courseId}:
    delete:
      description: Удаляет курс по ID
      parameters:
      - description: ID курса
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Удалить курс
      tags:
      - Courses
    get:
      description: Возвращает курс по его идентификатору
      parameters:
      - description: ID курса
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Получить курс по ID
      tags:
      - Courses
    put:
      consumes:
      - application/json
      description: Обновляет курс по ID
      parameters:
      - description: ID курса
        in: path
        name: courseId
        required: true
        type: string
      - description: Обновлённые данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Обновить курс
      tags:
      - Courses
...
swagger: "2.0"